Theming
Added styles.css file. It modifies the banner by creating a background box containing the dauphine-PSL logo and switching title color to white as well as centering it. It also modifies switch color font for date and section titles.

Modified _quarto.yml, changed the date and added the .css

Minor modification to Trade_lunch.qmd, added title-block-banner option to the yaml and small changes to the body of the text